Script to control Bone activity
The name of this script was originally correct, but the script evolved as the ball we were using for testing was replaced with the bone used in DinoFetch. This script contains some testing code, so I could test the game when I didn't have access to hardware and the final code to control the bone's behavior.