Droid Bullet Controller from VR multiplayer game
This is a script that controls the "bullets" coming from the droid in a multiplayer VR game that I worked on along with other classmates. My portion of the project was the droid, including this script. This defines all of the physics for the controller along with the networking of the bullets, using Photon.